+++ apps/fountain/lib/fountain_web/controllers/front_door_html.ex
@@ -1,0 +1,6 @@
+defmodule FountainWeb.FrontDoorHTML do
+  @moduledoc false
+  use FountainWeb, :html
+
+  embed_templates "front_door_html/*"
+end